Add files via upload
HQ is the "HAT" for a Raspberry Pi or compatible SBC.
DS is a most versatile "D"evice controller and "S"ensing module. It can be deployed to monitor voltages, currents and hence power and energy flowing through the device; it can control the flow like a switch including a PWM-based dimming function; it can also be used to sense a variety of analog as well as I2C-based inputs like temperature, pressure, humidity, fluid level, liquid flow and many other sensors.